chore(utils/__tests__): move bracket-conditional.test.ts from the bundler packages
